refactor column config into a new service
